requirements.txt
psycopg2~=2.9.1
discord.py~=1.7.1